requirements.txt
enaml>=0.10.2
twisted
enamlx
pyqtgraph
ipykernel<5; python_version < '3.0'
qtconsole
pyserial>=3.4
jsonpickle
lxml
faulthandler; python_version < '3.0'
PyQt5; python_version >= '3.0'
qt-reactor
pytest
pytest-coverage
pyusb
# for pytest-qt:
pytest-qt
codecov
